Adds middleman-vegas gem to handle syntax highlighting
Under the covers this uses the same rouge syntax lexer/highlighter
that was there before (albiet a later version). The gem provides
support to code fences and a code helper which is the same as
before but now it allows you to specify additional metadta after
the code language declaration for the code fence.

* Removes the javascript library that performed this highlighting
* Gemfile makes use of the new gem
* Config has been updated to load the gem by name
* A Rouge Lexer for Habitat Studios has been added to
  assist with processing the tokenization and ultimately
  its style and colorization
* Stylesheet has been added to support the new code format

@see https://github.com/burtlo/middleman-vegas#syntax

Signed-off-by: Franklin Webber <franklin@chef.io>