Fixes Slim Spacing
Ah this brings me back to taking ownership of a Python build system
with idention going about 10-space deep in a lot of places.
Whitespace can kill you. Even in small amounts.

I honestly don't know if these should all be flush or in one space
as they are... I'm opting to go with the flow.

Signed-off-by: Franklin Webber <franklin@chef.io>